Update task_controller.js
check for task befor deleting files in delete task function

+++ controller/task_controller.js
@@ -70,7 +70,8 @@ const updateTask = async (req,res)=>{
 const deleteTask = async (req,res)=>{
     try {
         const  task = await Task.findById(req.params.id.trim())
-        task.files.map(filename=>{
+        if(task){
+            task.files.map(filename=>{
             const filePath = path.join(__dirname.split('controller').join('uploads'),filename)
             fs.unlink(filePath,(err)=>{
                 if(err){
@@ -83,6 +84,7 @@ const deleteTask = async (req,res)=>{
             })
             
         })
+        }
         await Task.findByIdAndDelete(req.params.id.trim())
         res.status(200).json({
             status:"success",
